+++ Pulsus/Gameplay/Event/LongNoteEvent.cs
@@ -7,9 +7,9 @@ namespace Pulsus.Gameplay
 	{
 		public LongNoteEndEvent endNote;
 
-		public int length { get { return endNote.pulse - pulse; } }
+		public long length { get { return endNote.pulse - pulse; } }
 
-		public LongNoteEvent(int pulse, SoundObject sound, int lane, LongNoteEndEvent endNote)
+		public LongNoteEvent(long pulse, SoundObject sound, int lane, LongNoteEndEvent endNote)
 			: base(pulse, sound, lane)
 		{
 			this.endNote = endNote;
